[Gold V] Title: 입국심사, Time: 188 ms, Memory: 14764 KB -BaekjoonHub

+++ 백준/Gold/3079. 입국심사/입국심사.java
@@ -1,12 +1,12 @@
 public class Main {
 
-    static long n, m;
-    static long[] arr;
+    static int n, m;
+    static int[] arr;
 
     public static void main(String[] args) throws Exception {
         n = read();
         m = read();
-        arr = new long[(int) n];
+        arr = new int[n];
         long maxTime = 0;
         for (int i = 0; i < n; i++) {
             arr[i] = read();
@@ -27,13 +27,13 @@ public class Main {
         long cnt = 0;
         for (int i = 0; i < n; i++) {
             cnt += target / arr[i];
-            if(cnt >= m)
+            if (cnt >= m)
                 return true;
         }
         return cnt >= m;
     }
 
-    private static long read() throws Exception {
+    private static int read() throws Exception {
         int c, n = System.in.read() & 15;
         while ((c = System.in.read()) >= 48)
             n = (n << 3) + (n << 1) + (c & 15);
